kubeflow-trainer/pyproject.toml
[tool.poetry]
name = "kubeflow-trainer"
version = "0.1.0"
description = ""
authors = ["ya-mori <ppap.yohei.moriya@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
tensorflow-cpu = "^2.5.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
